requirements.txt
# Python=3.10.4
Django>=4.2.4,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.7,<2.10
djangorestframework-simplejwt>=5.3.0, <5.4
django-cors-headers>=4.2.0, <4.3
uWSGI>=2.0.22<2.1
drf-spectacular>=0.26.4,<1.0
drf-spectacular-sidecar>=2023.8.1,<2023.9